+++ lib/finance_rustler/solver.ex
@@ -7,7 +7,7 @@ defmodule FinanceRustler.Solver do
   solver at this module:
 
       # mix.exs
-      {:finance, "~> 1.5"},
+      {:finance, "~> 1.6"},
       {:finance_rustler, "~> 0.1"}
 
       # config/config.exs